requirements.txt
asgiref             3.2.7
asn1crypto          0.24.0
astroid             2.3.3
certifi             2018.8.24
cffi                1.14.0
chardet             3.0.4
click               6.7
clock               0.1
colorama            0.4.3
cryptography        2.9.2
distlib             0.3.0
Django              3.0.6
django-vote         2.2.0
et-xmlfile          1.0.1
filelock            3.0.12
Flask               1.0.2
Flask-Cors          3.0.6
Flask-Session       0.3.1
Flask-SocketIO      4.3.0
Flask-SQLAlchemy    2.3.2
idna                2.7
isort               4.3.21
itsdangerous        0.24
jdcal               1.4.1
Jinja2              2.10
lazy-object-proxy   1.4.3
MarkupSafe          1.0
mccabe              0.6.1
numpy               1.18.2
openpyxl            3.0.3
pandas              1.0.3
passlib             1.7.1
Pillow              7.1.2
pip                 20.1.1
psycopg2            2.8.5
psycopg2-binary     2.8.5
pycparser           2.19
pylint              2.4.4
pylint-django       2.0.15
pylint-plugin-utils 0.6
PyMySQL             0.9.2
python-dateutil     2.8.1
python-dotenv       0.13.0
python-engineio     3.12.1
python-socketio     4.5.1
pytz                2019.3
requests            2.19.1
setuptools          41.2.0
six                 1.15.0
SQLAlchemy          1.2.11
SQLAlchemy-ORM-tree 0.1.2
sqlparse            0.3.1
tzlocal             2.0.0
urllib3             1.23
virtualenv          20.0.20
Werkzeug            0.14.1
wrapt               1.11.2
xlrd                1.2.0